+++ src/rex.deploy/src/rex/deploy/column.py
@@ -13,6 +13,7 @@ from htsql.core.domain import (UntypedDomain, BooleanDomain, IntegerDomain,
 import datetime
 import decimal
 import json
+import collections
 
 
 class EnumValue(object):
@@ -239,6 +240,29 @@ class ColumnFact(Fact):
         if not self.is_present:
             args.append("is_present=%r" % self.is_present)
         return "%s(%s)" % (self.__class__.__name__, ", ".join(args))
+
+    def to_yaml(self, full=True):
+        mapping = collections.OrderedDict()
+        mapping['column'] = self.label
+        if full:
+            mapping['of'] = self.table_label
+        if self.former_labels:
+            mapping['was'] = self.former_labels
+        if full and self.front_labels:
+            mapping['after'] = self.front_labels
+        if self.type is not None:
+            mapping['type'] = self.type
+        if self.default is not None:
+            mapping['default'] = self.default
+        if self.is_required is False:
+            mapping['required'] = self.is_required
+        if self.is_unique is True:
+            mapping['unique'] = self.is_unique
+        if self.title is not None:
+            mapping['title'] = self.title
+        if self.is_present is False:
+            mapping['present'] = self.is_present
+        return mapping
 
     def __call__(self, driver):
         schema = model(driver)